requirements.txt
requests
imagehash
Pillow
argparse
